Destructured product
I would consider a change of the organic property on the product from organic to isOrganic for better practice because it is a boolean.  Also, I would rename the productName property to name as it is redundant as we already know that it is part of the product variable and you don't include product in front of any of the other properties, so it adds inconsistency.  Sorry for the run-on sentences, my teachers would kill me.